CentOS 7 glibc-ppc64le-46.patch: powerpc/fpu/libm-test-ulps tan() round-toward-zero fixup.
On POWER this patch also fixes test-ildoubl and test-ldouble failures where tan
rounded toward zero had acceptable 1 ULP error. Upstream is using 3 ULP, but
we prefer to keep the bound tighter unless we have a reason not to.

This is the ppc64le version which is required becuase it applies *after* another
ppc64le patch that touches the same ULPs file. See glibc-power-libm-test-ulps.patch
for the ppc64/ppc version.