CentOS 7 glibc-rh1165212.patch: This is a special patch for rhel-6 to fix recursive dlopen.
This is a special patch for rhel-6 to fix recursive dlopen.
It is likely the upstream patch will always be too risky for
rhel-6 and will involve reorganizing the way in which recursive
dlopen is allowed to operate and how the _r_debug and stap
points are used by gdb for the recursive case.

This fix changes the internal API to duplicate the ldconfig
cache data. This means that at any point the cache can be
unmapped without any consequences. The caller is responsible
fore freeing the returned string.

A regression test is added to verify the assertion for _r_debug
is no longer triggered due to the recursive dlopen. The test to
verify the fix in _dl_load_cache_lookup is not automated and
has to be run by hand.

The original version of this patch was based on the first version
of the upstream patch posted here:
  https://sourceware.org/ml/libc-alpha/2014-12/msg00446.html
The current version has been modified to reflect the changes
made in the revision of the patch committed to trunk after
being posted for review here:
  https://sourceware.org/ml/libc-alpha/2014-12/msg00483.html

This was committed upstream as:

commit ccdb048df457d581f6ac7ede8b0c7a593a891dfa
Author: Carlos O'Donell <carlos@systemhalted.org>
Date:   Wed Jan 21 01:51:10 2015 -0500

    Fix recursive dlopen.