CentOS 7 glibc-rh1227699.patch
This is the original fix for bug 1124987 in Fedora.
This patch should never go upstream. We are applying it
in rhel-7.2 as a workaround.

Upstream commit f8aeae347377f3dfa8cbadde057adf1827fb1d44 fixes
this issue correctly, but reveals other bugs, therefore for now
this patch is a workaround.