CentOS 7 glibc-rh1189278-1.patch: Also apply the RHEL6.7 Makerules patch (#1189278).
In rhel-6.x the Makerules are not entirely as mature as they are
in glibc 2.21 (from which the example link-libc-args is taken from).
In rhel-6.x the applications are not built like their counterpart
real applications, and because of that compiling DSOs that use TLS
will fail with undefined references to __tls_get_addr which resides
in ld.so and is never included in the link. This patch enhances
only the build-module and build-module-asneeded targets to include
a more fully and correct link line as the compiler driver would use
when constructing an application or DSO. We do not adjust the link
lines used to build lib* targets.