CentOS 7 glibc-rh1307241-1.patch: The upstream patch is backported by excluding a dependency in
The upstream patch is backported by excluding a dependency in
stdlib/Makefile whose task is to ensure that locales are generated before
the relevant tests are run. The dependency was added in order to fix
SWBZ#18969 which is triggered when running `make check' for a specific
subdirectory (stdlib here). Instead of this, the locales necessary for the
test are generated via an addition in localedata/Makefile.

commit 985fc132f23dbb83de76c5af9e783ef1b5900148
Author: Florian Weimer <fweimer@redhat.com>
Date:   Mon Apr 4 15:18:13 2016 +0200

    strfmon_l: Use specified locale for number formatting [BZ #19633]