CentOS 7 glibc-rh677316-RES_USE_INET6.patch: resolv: Deprecate the "inet6" option and RES_USE_INET6 [BZ #19582]
Backport this upstream commit, but without the resolv/resolv.h change
which introduces the deprecation warning:

commit b76e065991ec01299225d9da90a627ebe6c1ac97
Author: Florian Weimer <fweimer@redhat.com>
Date:   Tue Oct 4 11:52:10 2016 +0200

    resolv: Deprecate the "inet6" option and RES_USE_INET6 [BZ #19582]

No Makefile changes are needed here because we have not backported the
resolver testsuite at this point.  Furthermore, due to the lack of
actual RES_USE_INET6 deprecation, resolv/gethnamaddr.c is not modified
in this patch (because it still compiles).