CentOS 7 glibc-rh1418978-0.patch: backport upstream support/ directory
This patch backports the makefile changes related to the support/
framework.  It is based on upstream commits
c23de0aacbeaa7a091609b35764bed931475a16d (support: Introduce new
subdirectory for test infrastructure) and
76dcbf42df83c970c13c786d287f1ec69e1b91eb (Expose linking against
libsupport as make dependency).

The actual contents of the support/ subdirectory is kept in a separate
patch, so that it can be updated separately.