CentOS 7 glibc-rh1418978-max_align_t.patch: backport upstream support/ directory
This patch adds a definition of max_align_t to the internal
<sys/cdefs.h> header, for use in support/ functionality and elsewhere.

The definition must support being compiled in C11 or C++11 mode and
must check both __STDC_VERSION__ and __cplusplus versions.

Upstream does not need this because all files are compiled in C11
mode, making available the GCC definition in <stddef.h>

Also change malloc/tst-malloc-thread-fail.c to use the new definition.