CentOS 7 glibc-rh1401665-1a.patch: Additional robust mutex fixes
Upstream x86_64 uses the generic lll_futex_timed_wait_bitset, but for
rhel-7 we have not yet backported the larger generic futex cleanups.
To keep x86_64 using the tested code paths we add a generic
lll_futex_timed_wait_bitset.  We do not do the same for i686, where we
let the code use the lll_futex_timed_wait fallback.  For i686 it would
be harder to add lll_futex_timed_wait_bitset because it requires more
complex 6-argument syscall handling backports, so we do not do that.

Not needed upstream.