workaround for "left shift count >= width of type" warning by gcc
without this workaround, PIP-glibc cannot be compiled because of
the -Werror option of gcc.

NOTE: because the enterexit member is uint32_t, DL_NNS cannot exceed 32
with this logic.  So, I'm not resure whether this workaround is OK nor not.

this is part of the transition from
glibc-2.17-106.el7_2.8 RPM to
glibc-2.17-260.el7 RPM